+++ managelistings.py
@@ -9,11 +9,13 @@ import sqlite3
 class manageListings(QDialog):
     def __init__(self,app,uid):
         super(manageListings, self).__init__()
-        loadUi("managelist.ui", self)
+        loadUi("managelists.ui", self)
         self.confirmtoast.setText("")
         self.app = app
         self.userID = uid
         self.currentListingID = None
+        self.mode = None
+        self.activev = None
 
         self.loadTable()
         self.deletelist.clicked.connect(self.deletelisting)
@@ -34,20 +36,23 @@ class manageListings(QDialog):
             self.confirmtoast.setText("Select a \nRecord")
 
     def deletelisting(self):
+        self.mode = "Delete"
         self.fetchlistingID()
         self.confirmtoast.setText("Deleting \n listingID: " + str(self.currentListingID))
         self.confirmchoice.clicked.connect(self.deleterecord)
 
     def makelistingactive(self):
+        self.mode = "ChangeList"
         self.fetchlistingID()
         self.confirmtoast.setText("Making \n listingID: " + str(self.currentListingID) + "\nActive")
-        self.activev = 1
+        self.activev = True
         self.confirmchoice.clicked.connect(self.updatelisting)
 
     def makelistinginactive(self):
+        self.mode = "ChangeList"
         self.fetchlistingID()
         self.confirmtoast.setText("Making \n listingID: " + str(self.currentListingID) + "\nInactive")
-        self.activev = 0
+        self.activev = False
         self.confirmchoice.clicked.connect(self.updatelisting)
 
     def loadTable(self):
@@ -69,15 +74,17 @@ class manageListings(QDialog):
             header.setSectionResizeMode(i,QtWidgets.QHeaderView.Stretch)
 
     def deleterecord(self):
-        conn = sqlite3.connect("auc_database.db")
-        cur = conn.cursor()
-        cur.execute('DELETE FROM listings WHERE listingID = ?', (self.currentListingID,))
-        conn.commit()
-        self.loadTable()
+        if self.mode == "Delete":
+            conn = sqlite3.connect("auc_database.db")
+            cur = conn.cursor()
+            cur.execute('DELETE FROM listings WHERE listingID = ?', (self.currentListingID,))
+            conn.commit()
+            self.loadTable()
 
     def updatelisting(self):
-        conn = sqlite3.connect("auc_database.db")
-        cur = conn.cursor()
-        cur.execute('UPDATE listings SET active=? WHERE listingID=?', (self.activev, self.currentListingID))
-        conn.commit()
-        self.loadTable()
+        if self.mode == "ChangeList":
+            conn = sqlite3.connect("auc_database.db")
+            cur = conn.cursor()
+            cur.execute('UPDATE listings SET active=? WHERE listingID=?', (self.activev, self.currentListingID))
+            conn.commit()
+            self.loadTable()
